requirements.txt
streamlit
duckduckgo-search
beautifulsoup4
lxml
readability-lxml
requests
tqdm
pydantic
python-dotenv
numpy
scikit-learn
hdbscan
faiss-cpu
sentence-transformers